PerformanceApp/attempt3/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
python-multipart
pydantic-settings
pytest
httpx
